Main frontend work done
Most of the frontend is done. When the engine starts to get some functionality, the frontend will just have to call some some engine functions to check for move legality and stuff like that. I'm also sure that I'm saying this but will realize I want like 30 other things on the UI later but LETS GOOOOO we got this part in a working state